fix: schema.sql and propagte changes - renamed the last person in schema - added all the drop - propagated the changes throught the project (Serie, SerieRowMapper)
