+++ src/main/java/edu/lorsenmarek/backend/configuration/JDBCConfiguration.java
@@ -1,0 +1,4 @@
+package edu.lorsenmarek.backend.configuration;
+
+public class JdcbConfiguration {
+}
