+++ src/main/java/edu/lorsenmarek/backend/configuration/JacksonConfiguration.java
@@ -1,0 +1,2 @@
+package edu.lorsenmarek.backend.configuration;public class JacksonConfiguration {
+}
