+++ src/main/java/edu/lorsenmarek/backend/converter/jackson/InstantDeserializer.java
@@ -1,0 +1,29 @@
+package edu.lorsenmarek.backend.converter.jackson;
+
+import com.fasterxml.jackson.core.JacksonException;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+import java.time.Instant;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
+
+public class InstantDeserializer extends JsonSerializer<Instant>, JsonDeserializer<Instant> {
+    private static final DateTimeFormatter FORMATTER = DateTimeFormatter
+            .ofPattern("yyyy-MM-dd HH:mm:ss")
+            .withZone(ZoneOffset.UTC);
+    @Override
+    public void serialize(Instant instant, JsonGenerator jg, SerializerProvider serializerProvider) throws IOException {
+        jg.writeString(FORMATTER.format(instant));
+    }
+
+    @Override
+    public Instant deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException, JacksonException {
+        return null;
+    }
+}
