+++ src/main/java/edu/lorsenmarek/backend/converter/jackson/InstantSerializer.java
@@ -1,0 +1,4 @@
+package edu.lorsenmarek.backend.converter.jackson;
+
+public class InstantSerializer {
+}
