+++ src/main/java/edu/lorsenmarek/backend/converter/jdbc/InstantReadingConverter.java
@@ -1,0 +1,2 @@
+package edu.lorsenmarek.backend.converter.jdbc;public class InstantReadingConverter {
+}
