+++ src/main/java/edu/lorsenmarek/backend/converter/jdbc/SerieRowMapper.java
@@ -1,4 +1,4 @@
-package edu.lorsenmarek.backend.repository.mapper;
+package edu.lorsenmarek.backend.converter.jdbc;
 
 import edu.lorsenmarek.backend.model.Serie;
 import org.springframework.jdbc.core.RowMapper;
@@ -7,13 +7,14 @@ import org.springframework.lang.NonNull;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-public class SerieMapper implements RowMapper<Serie> {
+public class SerieRowMapper implements RowMapper<Serie> {
     @Override
     @NonNull
     public Serie mapRow(ResultSet rs, int rowNum) throws SQLException {
         return Serie.builder()
                 .id(rs.getLong("id"))
                 .title(rs.getString("title"))
+                .releasedAt(rs.getTimestamp("releasedAt").toInstant())
                 .build();
     }
 }
