+++ src/main/java/edu/lorsenmarek/backend/model/UserEpisodeRating.java
@@ -1,0 +1,2 @@
+package edu.lorsenmarek.backend.model;public class UserEpisodeRating {
+}
