+++ src/main/java/edu/lorsenmarek/backend/util/InstantCodecUtils.java
@@ -1,0 +1,5 @@
+package edu.lorsenmarek.backend.util;
+
+public class InstantCodecUtil {
+    
+}
